requirements.txt
Django==4.2.23
gunicorn==23.0.0
dj-database-url==3.0.1
psycopg2-binary==2.9.10
whitenoise==6.9.0

channels==4.2.2
channels_redis==4.2.1
djangorestframework==3.16.0
django-allauth==65.10.0
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
django-heroku==0.3.1

PyYAML
requests
redis==6.2.0